requirements.txt
python-dotenv
aiogram==2.25.1
ccxt
sqlalchemy
sqlmodel
cryptography
pydantic
matplotlib
scikit-learn
optuna
redis
rq
fastapi
uvicorn
pytest
yookassa
schedule
openai